debut de correction des bouts de map

+++ src/corewar/instructions/ft_lfork.c
@@ -30,7 +30,7 @@ void		ft_lfork(t_corevm *vm, t_process *process)
 	{
 		ft_printf("_____________________1create_process\n");
 		ft_printf("_____________________(process->pc + (process->args[0])) %d \n", (process->pc + (process->args[0])));
-		fork = create_process(vm, ((process->pc + (short)process->args[0]) & (MEM_SIZE - 1)), tmp);
+		fork = create_process(vm, ((process->pc + process->args[0]) & (MEM_SIZE - 1)), tmp);
 		ft_printf("_____________________fork->pc %d \n", fork->pc);
 		ft_printf("_____________________2create_process\n");
 	}
